chapter3_llm_evals/instructions/requirements.txt
streamlit-antd-components==0.2.5
streamlit
pytest
umap-learn
sentence-transformers
python-dotenv
plotly
pandas
openai==1.42.0
datasets
nbformat
jupyter
tabulate